backend/services/vector_store.py
import os
from dotenv import load_dotenv
from pinecone import Pinecone
from langchain_google_genai import GoogleGenerativeAIEmbeddings
from langchain_pinecone import PineconeVectorStore

# Load environment variables
load_dotenv()

# Initialize embedding model
embedding_model = GoogleGenerativeAIEmbeddings(model="models/embedding-001")

# Initialize Pinecone
pc = Pinecone(api_key=os.getenv("PINECONE_API_KEY"))
index = pc.Index("shop-product-catalog")

# Initialize vectorstore for PostgreSQL-based RAG
vectorstore = PineconeVectorStore(
    index=index,
    embedding=embedding_model,
    text_key="Description"
)