pyproject.toml
[project]
name = "rag-shop-assistant-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "google-generativeai==0.8.5",
    "kagglehub>=0.3.12",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "langchain-pinecone>=0.2.8",
    "pandas>=2.3.0",
    "pinecone-client>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.3",
]
